- Merged r31036-31046,31048,31050-31055,31061,31064,31066-31068,31070-31073,31076,31078-31079,31082-31087,31089-31092,31095-31101,31108-31118,31127-31128,31138-31139 from trunk.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2008/threadcomm@31155 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ data/shader/snippets/lighting/lightfuncs.cginc
@@ -25,7 +25,7 @@
 
 half Attenuation_Linear (float d, float invLightRadius)
 {
-  return saturate (1 - d * invLightRadius);
+  return (half)(saturate (1 - d * invLightRadius));
 }
 
 half Attenuation_Inverse (float d)
